Update resource limits and PVC creation in DITTO Kubeflow pipelines
This commit modifies the resource allocation settings in both `ditto_kubeflow_pipeline_gpu_enabled.py` and `ditto_kubeflow_pipeline.py`. The memory limit for GPU processing is adjusted to '16Gi', and the CPU limit is reduced to '4'. Additionally, the pipeline now creates a new persistent volume claim (PVC) dynamically instead of using an existing one, enhancing the flexibility of the pipeline setup. These changes aim to optimize resource usage and improve the overall efficiency of the DITTO entity matching process.