Enhance logging and update Hive parameters in DITTO Kubeflow pipelines
This commit introduces a logging mechanism in the `extract_hive_data_func` and `run_ditto_matching_func` to track execution details and errors. It also updates the Hive connection parameters in `ditto_kubeflow_pipeline_gpu_enabled.py` and `ditto_kubeflow_pipeline.py`, changing the user and database names for improved clarity. Additionally, a new function to create a summary of all logs from the pipeline run is added, enhancing the overall monitoring and debugging capabilities of the pipelines.