Update base image version and enforce image pull policy in DITTO Kubeflow pipeline
This commit updates the base image for the Kubeflow components in `ditto_kubeflow_pipeline_gpu_enabled.py` from version '1.5' to '2.0' to ensure compatibility with the latest features. Additionally, it sets the image pull policy to 'Always' for both the result processing and saving steps, ensuring that the most recent image is used during execution.